LPS-170489 Rename PortletConfigurationExtender for PortletConfigurationExtension and start PortletConfigurationExtension in LiferayServiceExtender.
